H. One Prime
H. One Prime
time limit per test3 seconds
memory limit per test64 megabytes
inputstandard input
outputstandard output
Given a number X. Determine if the number is prime or not

Note:

A prime number is a number that is greater than 1 and has only two factors which are 1 and itself.

In other words : prime number divisible only by 1 and itself.

Be careful that 1 is not prime .

The first few prime numbers are